Create turret model and materials
WIP
Its a pole with whats supposed to *very roughly* look like a cross between bolt action rifle and a chauchat welded to a slider joint.
The pole is connected to a metal plate which is bolted to a cement base.
The pole has a fast top that vaguely resembles a pickelhaube.